Update socket.js
These changes support transports to be one array which try websocket first, such as '[ 'websocket', 'polling']'.
The purpose is that we hope to give priority to websocket when implementing socket, in that polling will be blocked by the firewall instead of upgrading to websocket in some networked environments subject to security rules. But in ie9 - , it doesn’t support websocket, so we need to try the next protocol from the array of transports. The changes above are to achieve such a mechanism.