chore: downgrade the debug dependency
The newer versions need to be transpiled to ES5, so we'll rollback for
now.

We'll have to find a way to provide the users who bundle the Engine.IO
(or Socket.IO) client (with webpack for example) with a source code
which does not include debug (or document the usage of
webpack-remove-debug [1]).

Reverts https://github.com/socketio/engine.io-client/commit/242ea9edea8812724661e401a82a704670ed3bb1

[1] https://github.com/johngodley/webpack-remove-debug